refactor: further codebase organization
